+++ layers/GatoGraphQLForWP/plugins/gatographql/src/ModuleConfiguration.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace GatoGraphQL\GatoGraphQL;
 
-use GatoGraphQL\GatoGraphQL\Constants\LoggerSeverity;
 use PoP\Root\Module\AbstractModuleConfiguration;
 use PoP\Root\Module\EnvironmentValueHelpers;
 
@@ -174,40 +173,6 @@ class ModuleConfiguration extends AbstractModuleConfiguration
         $envVariable = Environment::ENABLE_SCHEMA_TUTORIAL_PAGE;
         $defaultValue = false;
         $callback = EnvironmentValueHelpers::toBool(...);
-
-        return $this->retrieveConfigurationValueOrUseDefault(
-            $envVariable,
-            $defaultValue,
-            $callback,
-        );
-    }
-
-    public function enableLogs(): bool
-    {
-        $envVariable = Environment::ENABLE_LOGS;
-        $defaultValue = false;
-        $callback = EnvironmentValueHelpers::toBool(...);
-
-        return $this->retrieveConfigurationValueOrUseDefault(
-            $envVariable,
-            $defaultValue,
-            $callback,
-        );
-    }
-
-    /**
-     * @return string[]
-     */
-    public function enableLogsBySeverity(): array
-    {
-        $envVariable = Environment::ENABLE_LOGS_BY_SEVERITY;
-        $defaultValue = [
-            LoggerSeverity::ERROR,
-            LoggerSeverity::WARNING,
-            LoggerSeverity::INFO,
-            LoggerSeverity::DEBUG,
-        ];
-        $callback = EnvironmentValueHelpers::commaSeparatedStringToArray(...);
 
         return $this->retrieveConfigurationValueOrUseDefault(
             $envVariable,