+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Overrides/Logger/Log/Logger.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GatoGraphQL\GatoGraphQL\Overrides\Logger\Log;
+
+use GatoGraphQL\GatoGraphQL\Log\Controllers\FileHandler\File;
+use PoPSchema\Logger\Log\Logger as UpstreamLogger;
+
+class Logger extends UpstreamLogger
+{
+    /**
+     * Generate the full name of a file based on source and date values.
+     *
+     * @param string $loggerSource The source property of a log entry, which determines the filename.
+     * @param int    $time   The time of the log entry as a Unix timestamp.
+     */
+    protected function generateLogFilename(string $loggerSource, int $time): string
+    {
+        $file_id = File::generate_file_id($loggerSource, null, $time);
+        $hash = File::generate_hash($file_id);
+
+        return "$file_id-$hash.log";
+    }
+}