+++ layers/Schema/packages/logger/src/Environment.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoPSchema\Logger;
+
+class Environment
+{
+    public final const LOGS_DIR = 'LOGS_DIR';
+    public final const ENABLE_LOGS = 'ENABLE_LOGS';
+    public final const ENABLE_LOGS_BY_SEVERITY = 'ENABLE_LOGS_BY_SEVERITY';
+}