+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -21,6 +21,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Improvements
 
 - Print the severity (`SUCCESS`/`INFO`/`WARNING`/`ERROR`) in the logs (#3099)
+- Use the Logger also when using plugin as standalone (without WordPress) (#3104)
 
 ## 12.2.2 - 13/05/2025